Fixed unit test cases
