+++ test/poi/test_poi.py
@@ -6,18 +6,16 @@ class TestPOI(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         self.poi = POI(
-            _id="0",
+            id="0",
             name="test",
-            clasification="TEST",
+            classification="TEST",
             hours_of_operation={"Monday": "11:00"},
             address="123 McMaster",
             poi_type="test",
             location={"longitude": "0", "latitude": "0"},
             image_url="picture",
         )
-
-    def test_to_dict(self):
-        poi_dict = {
+        self.poi_dict = {
             "_id": "0",
             "name": "test",
             "class": "TEST",
@@ -27,4 +25,9 @@ class TestPOI(unittest.TestCase):
             "location": {"longitude": "0", "latitude": "0"},
             "image_url": "picture",
         }
-        self.assertEqual(self.poi.to_dict(), poi_dict)
+
+    def test_to_dict(self):
+        self.assertEqual(self.poi.to_dict(), self.poi_dict)
+
+    def test_from_dict(self):
+        self.assertEqual(self.poi, POI.from_dict(self.poi_dict))
