+++ test/poi/test_poi_service.py
